스킬 해금 / 레벨업

+++ Mnamosyna/Assets/Scripts/_Monster/MonsterHPBar.cs
@@ -19,6 +19,6 @@ public class MonsterHPBar : MonoBehaviour
     
     void Update()
     {
-        hpImage.fillAmount = Monster.Cur_HP / Monster.Max_HP;
+        hpImage.fillAmount = (float)Monster.Cur_HP / Monster.Max_HP;
     }
 }
